+++ grid2op/Action/BaseAction.py
@@ -547,6 +547,24 @@ class BaseAction(GridObjects):
         self._modif_curtailment = False
         self._modif_alarm = False
 
+    def can_affect_something(self):
+        """
+        This functions returns True if the current action has any chance to change the grid.
+
+        Notes
+        -----
+        This does not say however if the action will indeed modify something somewhere !
+        """
+        return (self._modif_inj or
+                self._modif_set_bus or
+                self._modif_change_bus or
+                self._modif_set_status or
+                self._modif_change_status or
+                self._modif_redispatch or
+                self._modif_storage or
+                self._modif_curtailment or
+                self._modif_alarm)
+
     def _get_array_from_attr_name(self, attr_name):
         if hasattr(self, attr_name):
             res = super()._get_array_from_attr_name(attr_name)